move about into pages folder
