+++ src/App.js
@@ -1,7 +1,7 @@
 import { AppProvider } from "./components/AppContext";
 import Header from "./components/Header";
 import SongOverview from "./components/SongOverview";
-import About from "./components/About";
+import About from "./components/pages/About";
 import Footer from "./components/Footer";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
